Adding common abstract spell mechanic for all spell (having timer) NOT ALTERING EXISTING MECANICS - Tested with energyblast, to be tested with thor and witherskull
