+++ src/main/java/io/th0rgal/oraxen/mechanics/MechanicsManager.java
@@ -10,7 +10,7 @@ import io.th0rgal.oraxen.mechanics.provided.commands.CommandsMechanicFactory;
 import io.th0rgal.oraxen.mechanics.provided.consumable.ConsumableMechanicFactory;
 import io.th0rgal.oraxen.mechanics.provided.custom.CustomMechanicFactory;
 import io.th0rgal.oraxen.mechanics.provided.durability.DurabilityMechanicFactory;
-import io.th0rgal.oraxen.mechanics.provided.energyblast.EnergyBlastMechanicFactory;
+import io.th0rgal.oraxen.mechanics.provided.spell.energyblast.EnergyBlastMechanicFactory;
 import io.th0rgal.oraxen.mechanics.provided.harvesting.HarvestingMechanicFactory;
 import io.th0rgal.oraxen.mechanics.provided.hat.HatMechanicFactory;
 import io.th0rgal.oraxen.mechanics.provided.invisibleitemframe.InvisibleItemFrameFactory;
@@ -22,8 +22,8 @@ import io.th0rgal.oraxen.mechanics.provided.skin.SkinMechanicFactory;
 import io.th0rgal.oraxen.mechanics.provided.skinnable.SkinnableMechanicFactory;
 import io.th0rgal.oraxen.mechanics.provided.smelting.SmeltingMechanicFactory;
 import io.th0rgal.oraxen.mechanics.provided.soulbound.SoulBoundMechanicFactory;
-import io.th0rgal.oraxen.mechanics.provided.thor.ThorMechanicFactory;
-import io.th0rgal.oraxen.mechanics.provided.witherskull.WitherSkullMechanicFactory;
+import io.th0rgal.oraxen.mechanics.provided.spell.thor.ThorMechanicFactory;
+import io.th0rgal.oraxen.mechanics.provided.spell.witherskull.WitherSkullMechanicFactory;
 import io.th0rgal.oraxen.settings.ConfigUpdater;
 import io.th0rgal.oraxen.settings.ResourcesManager;
 import org.bukkit.Bukkit;
