+++ src/main/java/io/th0rgal/oraxen/mechanics/provided/spell/energyblast/EnergyBlastMechanicFactory.java
@@ -1,4 +1,4 @@
-package io.th0rgal.oraxen.mechanics.provided.energyblast;
+package io.th0rgal.oraxen.mechanics.provided.spell.energyblast;
 
 import io.th0rgal.oraxen.OraxenPlugin;
 import io.th0rgal.oraxen.mechanics.Mechanic;
