+++ src/main/java/io/th0rgal/oraxen/mechanics/provided/spell/thor/ThorMechanicFactory.java
@@ -1,0 +1,24 @@
+package io.th0rgal.oraxen.mechanics.provided.spell.thor;
+
+import io.th0rgal.oraxen.OraxenPlugin;
+import io.th0rgal.oraxen.mechanics.Mechanic;
+import io.th0rgal.oraxen.mechanics.MechanicFactory;
+import io.th0rgal.oraxen.mechanics.MechanicsManager;
+
+import org.bukkit.configuration.ConfigurationSection;
+
+public class ThorMechanicFactory extends MechanicFactory {
+
+    public ThorMechanicFactory(ConfigurationSection section) {
+        super(section);
+        MechanicsManager.registerListeners(OraxenPlugin.get(), new ThorMechanicListener(this));
+    }
+
+    @Override
+    public Mechanic parse(ConfigurationSection itemMechanicConfiguration) {
+        Mechanic mechanic = new ThorMechanic(this, itemMechanicConfiguration);
+        addToImplemented(mechanic);
+        return mechanic;
+    }
+
+}
