+++ src/main/java/io/th0rgal/oraxen/mechanics/provided/spell/witherskull/WitherSkullMechanic.java
@@ -1,13 +1,14 @@
-package io.th0rgal.oraxen.mechanics.provided.witherskull;
+package io.th0rgal.oraxen.mechanics.provided.spell.witherskull;
 
 import io.th0rgal.oraxen.mechanics.Mechanic;
 import io.th0rgal.oraxen.mechanics.MechanicFactory;
+import io.th0rgal.oraxen.mechanics.provided.spell.SpellMechanic;
 import io.th0rgal.oraxen.utils.timers.Timer;
 import io.th0rgal.oraxen.utils.timers.TimersFactory;
 import org.bukkit.configuration.ConfigurationSection;
 import org.bukkit.entity.Player;
 
-public class WitherSkullMechanic extends Mechanic {
+public class WitherSkullMechanic extends SpellMechanic {
 
     private final TimersFactory timersFactory;
     public final boolean charged;
